Moved template CheckSum code into header file to make usable outside ShipDesign.h.
Conflicts:
	universe/ShipDesign.cpp
